requirements.txt
transformers
accelerate
lm_eval
tqdm
deepspeed
wandb
hqq
optimum-quanto
torchvision
lightning>=2.0.0
pytorch-lightning>=2.0.0
pydantic
xformers
flash-attn==2.8.2
bitsandbytes
datasets
sentencepiece
jsonargparse
hydra-core>=1.3.0
omegaconf>=2.3.0
